requirements.txt
# Core
Django>=4.2,<5.0
python-decouple>=3.8

# AI
google-generativeai>=0.5.0

# Database (production)
psycopg2-binary>=2.9.9

# Testing
pytest>=7.4
pytest-django>=4.7
pytest-cov>=4.1
factory-boy>=3.3

# Code quality
black>=24.0
isort>=5.13
flake8>=7.0

# Production
gunicorn>=21.2
whitenoise>=6.6
sentry-sdk>=1.40

# Utilities
python-dateutil>=2.8
